Module icaros_control::gamepad

Gamepad support

Structs

ControlState
Gamepad state, shared with the main system

PhysicalGamepadAxesCodes
Holds the driver identifiers for each control (or "codes").

PhysicalGamepadService
The physical gamepad input service, which takes inputs from a physical gamepad.

RawAxisState
Raw processed axis state.

Constants

THROTTLE_CHANGE_SPEED
A constant factor that determines the rate to which the throttle changes when input is provided.

THROTTLE_QUICK_TRIGGER_BOUND
A constant factor that determines how much a full press of one of the triggers may influence the final throttle, in percentage terms (i.e. 0.1 means it may influence the throttle up or down up to a 10% of the total value range).

_IMPL_DESERIALIZE_FOR_ControlState

_IMPL_DESERIALIZE_FOR_RawAxisState

_IMPL_SERIALIZE_FOR_ControlState

_IMPL_SERIALIZE_FOR_RawAxisState

Traits

GamepadService
A gamepad input service.