Trait icaros_control::gamepad::GamepadService

pub trait GamepadService {
    fn is_controller_available(&mut self) -> bool;

    fn process(
        &mut self, 
        state: &mut ControlState
    ) -> Result<bool, Box<dyn Error>>;

    fn control_state_into_message_state(
        &mut self, 
        msg_state: &mut G2AControllerAxisState, 
        ctrl_state: &ControlState
    );
}

A gamepad input service.

Implementors of this trait can act as services that read user input on the drone's state from a device such as a gamepad and pass it to the main system for routing into the drone and/or appropiate subsystems.

Required methods

fn is_controller_available(&mut self) -> bool

Returns whether the controller is available for reading

fn process(&mut self, state: &mut ControlState) -> Result<bool, Box<dyn Error>>

Read the gamepad inputs and update the given state with them

fn control_state_into_message_state(
&mut self,
msg_state: &mut G2AControllerAxisState,
ctrl_state: &ControlState
)

Convert the given state into a processed, final control state ready to send to the drone

Loading content...