Module icaros_control::web

Main webserver thread logic

Unlike the rest of the code, this thread runs on the Tokio asyncronomous event loop for better performance and semantics for the task.

To handle the web side itself, the warp web server library is used. Three endpoints are exposed from it:

/feed, which represents our core WebSocket endpoint, and streams the main system state and telemetry to connected users.

/video, which represents our video WebSocket endpoint, and streams the live video feed received from the drone through WifiBroadcast.

A catch-all endpoint, which serves the front-end side of the web application.

Modules

video
Video processing for the web side

ws_events
Websocket common event definitions (i.e. connect, disconnect)

Structs

WebserverSharedData
The complete state of the webserver

WebsocketEndpoint
The complete state of a single WebSocket endpoint.

Websockets
Contains the two main WebSocket endpoints, core and video.

Enums

ControllerRXWebserverMessage
Kinds of signals the main system can send to the webserver

ControllerTXWebserverMessage
Kinds of signals the webserver can send to the main system

Functions

webserver_main
Entry point for the web server thread.

Type Definitions

Users
A bare-bones key-value store of WebSocket sockets, mapped by user ID

WebsocketSender
Message-passing interface for WebSocket messages behind which is the WebSocket socket Messages for WebSocket clients are written through here.