Module icaros_control::web::video

Video processing for the web side

Background

A H.264 video stream is composed of NAL units (NALU). Each NAL unit starts with [0x00, 0x00, 0x00, 0x01]1. Our only objective is to split these unit using this delimiter to get the individual NAL units, which the web server can then send to its clients.

https://yumichan.net/video-processing/video-compression/introduction-to-h264-nal-unit/ ↩

Structs

NALSeparator
A decoder which takes a H.264 bitstream and outputs a series of framed bitstreams corresponding to the individual NAL units extracted from it.

NAL_SEPARATOR_REGEX
A regular expression which can find the byte combination that delimits H.264 NAL units on a byte slice.

Enums

VideoControlPayload
The different payloads that comphrend the basic protocol used by the front-end to configure the stream.

Constants

NAL_SEPARATOR
The byte combination that delimits H.264 NAL units.

_IMPL_SERIALIZE_FOR_VideoControlPayload

Functions

separate_nal
Creates a decoder which which takes a H.264 bitstream and outputs a series of framed bitstreams corresponding to the individual NAL units extracted from it.